refactor: changes in statistics
- все же среднее значение целых чисел может быть не целым числом
- все же вероятность найти строку из >2_147_483_647 символов
пренебрежимо мала

+++ app/src/main/java/ru/izedikus/filefilter/models/Statistics.java
@@ -11,13 +11,13 @@ public record Statistics(
   BigInteger sumInt,
   BigInteger minInt,
   BigInteger maxInt,
-  BigInteger avgInt,
+  BigDecimal avgInt,
 
   BigDecimal sumFloat,
   BigDecimal minFloat,
   BigDecimal maxFloat,
   BigDecimal avgFloat,
 
-  Long minStringLen,
-  Long maxStringLen
+  int minStringLen,
+  int maxStringLen
   ) {}
